src/requirements.txt
# Web framework
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# AI/ML dependencies
langchain>=0.0.350
langchain-community>=0.0.10
crewai>=0.1.0
autogen>=0.2.0
llama-index>=0.9.0
qdrant-client>=1.7.0

# Audio processing
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# Video processing
opencv-python>=4.8.0
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.1.0
tenacity>=8.2.0
tqdm>=4.66.1
requests>=2.31.0
aiohttp>=3.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Additional dependencies
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
ray[serve]>=2.7.0
mlflow>=2.8.0
prometheus-client>=0.17.1
aiofiles>=23.2.1
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.14.0
wandb>=0.15.12 